feat: upgrade bandit reward accumulation to f64 precision
- Update `RemindBandit` in `heimlern-bandits` to use `HashMap<String, (u32, f64)>` for internal value storage, reducing floating-point drift over time.
- Update `ContractSnapshot` to serialize values as `Vec<f64>`, preserving precision during persistence.
- Update `OutcomeStatistics` in `heimlern-feedback` to use `f64` for `total_reward`.
- Maintain API compatibility by casting to `f32` in `Policy` trait methods and public accessors where required.
- Update tests to use `f64` literals (e.g. `f64::INFINITY`).

+++ crates/heimlern-feedback/src/lib.rs
@@ -160,7 +160,7 @@ pub struct OutcomeStatistics {
     pub total: usize,
     pub successes: usize,
     pub failures: usize,
-    pub total_reward: f32,
+    pub total_reward: f64,
 }
 
 impl OutcomeStatistics {
@@ -197,7 +197,7 @@ impl OutcomeStatistics {
         }
         #[allow(clippy::cast_precision_loss)]
         {
-            self.total_reward / self.total as f32
+            (self.total_reward / self.total as f64) as f32
         }
     }
 }
@@ -250,7 +250,7 @@ impl FeedbackAnalyzer {
                 }
                 if let Some(reward) = outcome.reward {
                     if reward.is_finite() {
-                        entry.total_reward += reward;
+                        entry.total_reward += f64::from(reward);
                     }
                 }
             }
@@ -271,7 +271,7 @@ impl FeedbackAnalyzer {
             }
             if let Some(reward) = outcome.reward {
                 if reward.is_finite() {
-                    stats.total_reward += reward;
+                    stats.total_reward += f64::from(reward);
                 }
             }
         }
